New Terrain Generator

+++ include/Shape.h
@@ -6,7 +6,7 @@ struct TransformationComponent;
 
 struct Triangle {
 	glm::vec3 p1, p2, p3;
-	glm::vec2 t1, t2, t3;
+	glm::ivec2 t1, t2, t3;
 	glm::vec3 normal;
 };
